feat: tasks 1-7 done
